centralize even more code (email address getter for keycloak users)
