+++ dataland-email-service/src/main/kotlin/org/dataland/datalandemailservice/services/TemplateEmailMessageListener.kt
@@ -2,6 +2,8 @@ package org.dataland.datalandemailservice.services
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
+import okhttp3.OkHttpClient
+import org.dataland.datalandbackendutils.utils.getEmailAddress
 import org.dataland.datalandemailservice.email.EmailSender
 import org.dataland.datalandemailservice.services.templateemail.TemplateEmailFactory
 import org.dataland.datalandmessagequeueutils.constants.ExchangeName
@@ -17,6 +19,8 @@ import org.springframework.amqp.rabbit.annotation.Queue
 import org.springframework.amqp.rabbit.annotation.QueueBinding
 import org.springframework.amqp.rabbit.annotation.RabbitListener
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Qualifier
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.messaging.handler.annotation.Header
 import org.springframework.messaging.handler.annotation.Payload
 import org.springframework.stereotype.Service
@@ -31,7 +35,8 @@ class TemplateEmailMessageListener(
     @Autowired private val messageQueueUtils: MessageQueueUtils,
     @Autowired private val objectMapper: ObjectMapper,
     @Autowired private val templateEmailFactories: List<TemplateEmailFactory>,
-    @Autowired private val keycloakUserControllerApiService: KeycloakUserControllerApiService,
+    @Qualifier("AuthenticatedOkHttpClient") val authenticatedOkHttpClient: OkHttpClient,
+    @Value("\${dataland.keycloak.base-url}") private val keycloakBaseUrl: String,
 ) {
     private val logger = LoggerFactory.getLogger(TemplateEmailMessageListener::class.java)
 
@@ -84,7 +89,12 @@ class TemplateEmailMessageListener(
     private fun getEmailAddressByRecipient(receiver: JsonNode?): String {
         return when (val receiverType = receiver?.get("type")?.asText()) {
             "address" -> receiver.get("email").asText()
-            "user" -> keycloakUserControllerApiService.getEmailAddress(receiver.get("userId").asText())
+            "user" -> getEmailAddress(
+                authenticatedOkHttpClient,
+                objectMapper,
+                keycloakBaseUrl,
+                receiver.get("userId").asText(),
+            )
             else -> {
                 throw IllegalArgumentException("Invalid receiver type: $receiverType")
             }
